Lung-Cancer-Classification/requirements.txt
flask==3.1.0
tensorflow==2.18.0
gunicorn==23.0.0
numpy
pillow
waitress